+++ changelog.md
@@ -5,6 +5,10 @@
 - added rule `habapp_rules.actors.shading.ReferenceRun` to trigger the reference run for blinds every month
 - added rule `habapp_rules.system.task.RecurringTask` to trigger recurring tasks
 
+# Bugfix
+
+- removed timezone from all datetime.items, since timezone makes no sense in the OpenHAB context
+
 # Version 7.1.0 - 01.12.2024
 
 # Features